requirements.txt
asgiref==3.5.2
autopep8==1.6.0
cffi==1.15.0
cryptography==37.0.2
dj-database-url==0.5.0
Django==4.0.4
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.1.1
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
whitenoise==6.1.0
